requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
requests==2.31.0
Pillow==10.1.0
email-validator==2.1.0.post1
PyJWT==2.8.0
redis==5.0.1
boto3==1.28.64
gunicorn==21.2.0 